- 16 September 2023 Update

+++ mysite/render.py
@@ -6,6 +6,7 @@ from django.http import HttpResponse
 from django.template.loader import get_template
 
 from .s3_file_manager import S3
+from .settings import ENV
 
 
 class Render:
@@ -32,17 +33,19 @@ class Render:
             os.makedirs(os.path.join(os.path.abspath(os.path.dirname("__file__")), "media/pdfs/" + pdf_type))
         with open(file_path, 'wb') as pdf:
             pisa.pisaDocument(BytesIO(html.encode("UTF-8")), pdf)
-        s3 = S3()
-        s3.upload_file_to_bucket(file_name=file_path, key=s3_path)
-        if delete_file:
-            os.remove(file_path)
-        file_path = s3.get_bucket_object(s3_path)
+        if ENV != 'local':
+            s3 = S3()
+            s3.upload_file_to_bucket(file_name=file_path, key=s3_path)
+            if delete_file:
+                os.remove(file_path)
+            file_path = s3.get_bucket_object(s3_path)
         return [file_name, file_path]
 
     @staticmethod
     def delete_file(params: dict, pdf_type: str):
         file_name = "{0}.pdf".format(params.get('file_name'))
         file_path = "media/pdfs/" + pdf_type + '/' + file_name
-        s3 = S3()
-        s3.delete_file_from_bucket(key=file_path)
+        if ENV != 'local':
+            s3 = S3()
+            s3.delete_file_from_bucket(key=file_path)
         return 1
